refactor: extract UVToolMixin, cache find_executable, consolidate utils
- Extract 13 UV tool methods from PluginBase (1174 -> 355 lines) into
  base/uv_tool_mixin.py.
- Extract _process_command_response to deduplicate response handling.
- Add find_executable caching with invalidation after tool install.
- Add default validate_plugin_dependencies in ScannerPluginBase.
- Move path_matches_pattern from sarif_utils to normalizers.
- Consolidate get_ash_version to delegate to version_management.
- Single os.walk pass in get_scan_set.py (was triple walk).
- Remove unused import os from sarif_utils, plugin_base.
