+++ automated_security_helper/utils/get_ash_version.py
@@ -1,8 +1,8 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-from importlib.metadata import version
-
 
 def get_ash_version():
-    return version("automated_security_helper")
+    from automated_security_helper.utils.version_management import get_version
+
+    return get_version()
